requirements.txt
aiohttp==3.8.1
aiohttp-retry==2.4.6
aiosignal==1.2.0
alembic==1.8.0
appdirs==1.4.4
async-timeout==4.0.2
asyncssh==2.11.0
atpublic==3.0.1
attrs==21.4.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cloudpickle==2.1.0
colorama==0.4.4
commonmark==0.9.1
configobj==5.0.6
cryptography==37.0.2
cycler==0.11.0
databricks-cli==0.16.6
dictdiffer==0.9.0
diskcache==5.4.0
distro==1.7.0
docker==4.2.2
dpath==2.0.6
dulwich==0.20.43
dvc==2.11.0
dvc-data==0.0.2
dvc-objects==0.0.2
dvc-render==0.0.5
dvclive==0.8.2
entrypoints==0.4
Flask==2.1.2
flatten-dict==0.4.2
flufl.lock==7.0
fonttools==4.33.3
frozenlist==1.3.0
fsspec==2022.5.0
ftfy==6.1.1
funcy==1.17
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
grandalf==0.6
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.2
mailchecker==4.1.17
Mako==1.2.0
MarkupSafe==2.1.1
matplotlib==3.5.2
mlflow==1.26.1
multidict==6.0.2
nanotime==0.5.2
networkx==2.8.2
numpy==1.22.4
oauthlib==3.2.0
packaging==21.3
pandas==1.4.2
pathspec==0.9.0
phonenumbers==8.12.49
Pillow==9.1.1
prometheus-client==0.14.1
prometheus-flask-exporter==0.20.2
protobuf==4.21.1
psutil==5.9.1
pycparser==2.21
pydot==1.4.2
pygit2==1.9.2
Pygments==2.12.0
pygtrie==2.4.2
PyJWT==2.4.0
pyparsing==3.0.9
pypiwin32==223
python-benedict==0.25.1
python-dateutil==2.8.2
python-fsutil==0.6.1
python-slugify==6.1.2
pytz==2022.1
pywin32==304
PyYAML==6.0
querystring-parser==1.2.4
requests==2.27.1
rich==12.4.4
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.1.1
scipy==1.8.1
scmrepo==0.0.24
shortuuid==1.0.9
shtab==1.5.4
six==1.16.0
sklearn==0.0
smmap==5.0.0
SQLAlchemy==1.4.37
sqlparse==0.4.2
tabulate==0.8.9
text-unidecode==1.3
threadpoolctl==3.1.0
toml==0.10.2
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
voluptuous==0.13.1
waitress==2.1.2
wcwidth==0.2.5
websocket-client==1.3.2
Werkzeug==2.1.2
xmltodict==0.13.0
yarl==1.7.2
zc.lockfile==2.0
zipp==3.8.0
